refactor: split stores into modules
