+++ components/LoginDialog.tsx
@@ -8,7 +8,7 @@ import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Checkbox } from "@/components/ui/checkbox"
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from "@/components/ui/dialog"
-import { useAuthStore } from "@/lib/store"
+import { useAuthStore } from "@/lib/stores"
 import { supabaseBrowser } from "@/lib/supabase"
 import { useToast } from "@/hooks/use-toast"
