+++ lib/stores/index.ts
@@ -1,0 +1,5 @@
+export { useAuthStore } from './auth';
+export { useProductStore } from './product';
+export { useLeadStore } from './lead';
+export { useCartStore } from './cart';
+export type { CartItem } from './cart';
