+++ lib/stores/product.ts
@@ -1,0 +1,37 @@
+import { create } from "zustand";
+import { persist } from "zustand/middleware";
+import { Product } from "@/types/product";
+
+interface ProductState {
+  products: Product[];
+  addProduct: (product: Omit<Product, "id">) => void;
+  updateProduct: (id: string, product: Partial<Product>) => void;
+  deleteProduct: (id: string) => void;
+  getProduct: (id: string) => Product | undefined;
+}
+
+export const useProductStore = create<ProductState>()(
+  persist(
+    (set, get) => ({
+      products: [],
+      addProduct: (product) =>
+        set((state) => ({
+          products: [...state.products, { ...product, id: Date.now().toString() }],
+        })),
+      updateProduct: (id, updatedProduct) =>
+        set((state) => ({
+          products: state.products.map((product) =>
+            product.id === id ? { ...product, ...updatedProduct } : product
+          ),
+        })),
+      deleteProduct: (id) =>
+        set((state) => ({
+          products: state.products.filter((product) => product.id !== id),
+        })),
+      getProduct: (id) => get().products.find((product) => product.id === id),
+    }),
+    {
+      name: "product-store",
+    }
+  )
+);
